WIP nimble now builds
